LottoGenerator 추상화를 위해 기존 LottoGenerator -> AutoLottoGenerator로 변경 및 LottoGenerator interface 생성
