+++ src/main/java/lotto/domain/LottoGenerator.java
@@ -1,45 +1,5 @@
 package lotto.domain;
 
-import java.util.Collections;
-import java.util.Comparator;
-import java.util.List;
-import java.util.stream.Collectors;
-import java.util.stream.IntStream;
-
-public class LottoGenerator {
-    private LottoGenerator() {
-    }
-
-    public static Lotto generate() {
-        return new Lotto(getSortedLottoNumbers());
-    }
-
-    private static List<LottoNumber> getSortedLottoNumbers() {
-        List<LottoNumber> lottoNumbers = getUnsortedLottoNumbers();
-        sort(lottoNumbers);
-
-        return lottoNumbers;
-    }
-
-    private static void sort(List<LottoNumber> lottoNumbers) {
-        lottoNumbers.sort(Comparator.naturalOrder());
-    }
-
-    private static List<LottoNumber> getUnsortedLottoNumbers() {
-        return getShuffledAllLottoNumbers().subList(0, Lotto.LOTTO_NUMBERS_SIZE);
-    }
-
-    static List<LottoNumber> getAllLottoNumbers() {
-        return IntStream.rangeClosed(LottoNumber.MIN_NUMBER, LottoNumber.MAX_NUMBER)
-                .boxed()
-                .map(LottoNumber::getInstance)
-                .collect(Collectors.toList());
-    }
-
-    private static List<LottoNumber> getShuffledAllLottoNumbers() {
-        List<LottoNumber> allLottoNumbers = getAllLottoNumbers();
-        Collections.shuffle(allLottoNumbers);
-
-        return allLottoNumbers;
-    }
+public interface LottoGenerator {
+    Lotto generate();
 }
